+++ CHANGES.rst
@@ -67,6 +67,7 @@ Development
 - Add backup mechanism for configuration files
 - Fix polling for WiFi connectivity
 - Add basic HTTP API endpoints for configuration settings
+- Add backup configuration snippet to settings blueprint files
 
 
 2019-06-22 0.5.1